backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_restful
flask_marshmallow
marshmallow_sqlalchemy